:heartpulse: Extract request creation and URL processing
