+++ lib/buildResponse.js
