+++ lib/constants.js
@@ -1,3 +1,4 @@
 module.exports = {
+  validDomain: 'nhs.uk',
   validEnvironments: ['staging', 'production'],
 };
