+++ lib/processURLs.js
@@ -1,0 +1,29 @@
+const { validDomain } = require('../lib/constants');
+
+function processURLs(objects) {
+  const urls = objects.filter(Boolean).map((url) => url.trim());
+  const uniqueURLs = [...new Set(urls)];
+  const unparseableURLs = [];
+  const invalidURLs = [];
+
+  uniqueURLs.forEach((url) => {
+    try {
+      const parsedURL = new URL(url);
+      if (!parsedURL.host.endsWith(validDomain)) {
+        invalidURLs.push(url);
+      }
+    } catch (err) {
+      unparseableURLs.push(url);
+    }
+  });
+
+  return {
+    invalidURLs,
+    uniqueURLs,
+    unparseableURLs,
+  };
+}
+
+module.exports = {
+  processURLs,
+};
